catalog_service/requirements.txt
fastapi
uvicorn
sqlalchemy
sqlalchemy[asyncio]
asyncpg
python-dotenv
pydantic
python-jose

# Tests
pytest
pytest-asyncio
httpx
aiosqlite
alembic
